Break out product tests, introduce data providers
